+++ lib/src/class/artist/artist.dart
@@ -8,4 +8,9 @@ class Artist {
   }
   String name;
   List<Album> albums;
+
+  @override
+  String toString() {
+    return "Name: " + this.name + " has " + albums.length.toString() + " albums\n";
+  }
 }