+++ lib/src/class/songinfo/songinfo.dart
@@ -4,8 +4,7 @@ import 'package:playify/src/class/artist/artist.dart';
 import 'package:playify/src/class/song/song.dart';
 
 class SongInfo {
-  SongInfo(
-      {@required Album album, @required Song song, @required Artist artist}) {
+  SongInfo({@required Album album, @required Song song, @required Artist artist}) {
     this.album = album;
     this.song = song;
     this.artist = artist;
@@ -13,4 +12,14 @@ class SongInfo {
   Album album;
   Song song;
   Artist artist;
+
+  @override
+  String toString() {
+    return "Artist: " +
+        this.artist.toString() +
+        "Album: " +
+        this.album.toString() +
+        "Song: " +
+        this.song.toString();
+  }
 }